modularize post
